Fix:컨플릭트 수정

+++ src/main/java/com/TMT/TMT_BE_PaymentServer/wallet/dto/SendWalletInfoDto.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class SendWalletInfoDto {
 
     private String uuid;
-    private int won;
+    private Long won;
 
 }
